Config changes: All config including login is all one file setup.
Login Changes:
Cookies should be properly used now.
Login module should be fine except for once loging in and changing pages the imput form shows back up.
Visitor.php is used when a visitor tries to open a members only page. can be changed to whatever u want it to be.
Teleporter-test.php added with session functions but i cannot for the life of me to get it to work with the login info.

Navigation changes:
Forgot to add change pass and get pass to the leftnavi.php

Getting closer and closer :P